COMMERCE-481 Pull up the getters for the common terms
