+++ modules/etl/talend/talend-runtime/src/main/java/com/liferay/talend/runtime/apio/jsonld/ApioResponse.java
@@ -23,7 +23,11 @@ public interface ApioResponse {
 
 	public JsonNode getContextJsonNode();
 
+	public String getDescription();
+
 	public JsonNode getIdJsonNode();
+
+	public String getTitle();
 
 	public JsonNode getTypeJsonNode();
